Reload client in case of "No Match Error" (#8)
* Reload client in case of "No Match Error"
